% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lum_tbl}
\alias{lum_tbl}
\title{Luminx response data}
\format{
\subsection{\code{who}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
lum_tbl
}
\description{
Data containing responses to Mtb antigens from the Luminex assay.
}
\keyword{datasets}
